Folds single_thread_sampler.* into gibbs_sampler.{cc,h}
