+++ Makefile
@@ -53,7 +53,6 @@ SOURCES += src/factor.cc
 SOURCES += src/factor_graph.cc
 SOURCES += src/inference_result.cc
 SOURCES += src/gibbs_sampling.cc
-SOURCES += src/single_thread_sampler.cc
 SOURCES += src/gibbs_sampler.cc
 SOURCES += src/timer.cc
 OBJECTS = $(SOURCES:.cc=.o)
